pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Azalurg",email = "91213389+Azalurg@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.15.1,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uuid (>=1.30,<2.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "numpy (>=2.2.3,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.4"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100 
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
